requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
catboost>=1.2.0
grad-cam>=1.4.0  # Installs as pytorch_grad_cam
mahotas>=1.4.0
scikit-learn>=1.3.0
joblib>=1.3.0
lime>=0.2.0.1
scikit-image>=0.21.0

